+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## XX.XX.XX
+* Extended to allow custom handling of URLs when notification buttons are clicked in the background.
+
 ## 25.4.0
 * ! Minor breaking change ! Removed Secure.ANDROID_ID usage in device id generation. The SDK now exclusively uses random UUIDs for device id generation.
 * ! Minor breaking change ! Server Configuration is now enabled by default. Changes made on SDK Manager > SDK Configuration on your server will affect SDK behavior directly.